Actualización deepSpaceConquerors

+++ src/Carta.java
@@ -2,9 +2,9 @@
 public abstract class Carta {
 
     String nombreCarta;
-    double precio;
+    int precio;
 
-    public Carta(String nombreCarta, double precio) {
+    public Carta(String nombreCarta, int precio) {
         this.nombreCarta = nombreCarta;
         this.precio = precio;
     }
